--- a/package.json
+++ b/package.json
@@ -73,6 +73,7 @@
     "@release-it-plugins/lerna-changelog": "^6.0.0",
     "@release-it-plugins/workspaces": "^4.0.0",
     "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-strip": "^3.0.4",
     "@types/babel-plugin-macros": "^3.1.3",
     "@types/babel__core": "^7.20.5",
     "@types/babel__traverse": "^7.20.4",
@@ -103,8 +104,6 @@
     "eslint-plugin-unused-imports": "^3.0.0",
     "esyes": "^1.0.1",
     "execa": "^7.1.1",
-    "tracerbench": "^8.0.1",
-    "zx": "^7.2.3",
     "fast-glob": "^3.2.12",
     "glob": "^10.2.3",
     "js-yaml": "^4.1.0",
@@ -122,11 +121,13 @@
     "semver": "^7.5.2",
     "testem-failure-only-reporter": "^1.0.0",
     "toml": "^3.0.0",
+    "tracerbench": "^8.0.1",
     "ts-node": "^10.9.1",
     "turbo": "^1.9.3",
     "typescript": "^5.0.4",
     "vite": "^5.0.10",
-    "xo": "^0.54.2"
+    "xo": "^0.54.2",
+    "zx": "^7.2.3"
   },
   "release-it": {
     "plugins": {