--- a/packages/@glimmer/util/index.ts
+++ b/packages/@glimmer/util/index.ts
@@ -6,7 +6,6 @@ export { default as debugToString } from './lib/debug-to-string';
 export * from './lib/dom';
 export * from './lib/dom-utils';
 export * from './lib/immediate';
-export { default as intern } from './lib/intern';
 export {
   isSerializationFirstNode,
   SERIALIZATION_FIRST_NODE_STRING,