--- a/packages/@glimmer/validator/index.ts
+++ b/packages/@glimmer/validator/index.ts
@@ -44,6 +44,7 @@ export {
   DIRTY_TAG as dirtyTag,
   INITIAL,
   isConstTag,
+  prepareTagsIfNeeded,
   type Revision,
   UPDATE_TAG as updateTag,
   validateTag,