… packages/@glimmer/util/lib/intern.ts
… pnpm-lock.yaml