[[perf] improve list retaining logic](https://github.com/glimmerjs/glimmer-vm/pull/1531)
[[perf] tags pool experiment](https://github.com/glimmerjs/glimmer-vm/pull/1521)
[[perf] remove extra combine logic](https://github.com/glimmerjs/glimmer-vm/pull/1515)
[[chore] simplify few opcodes](https://github.com/glimmerjs/glimmer-vm/pull/1519)
[[chore] test prod terser config](https://github.com/glimmerjs/glimmer-vm/pull/1536)
[[chore] remove ie11 on fallback](https://github.com/glimmerjs/glimmer-vm/pull/1535)

<img width="796" alt="image" src="https://github.com/glimmerjs/glimmer-vm/assets/1360552/87d9e09f-0475-4dea-ada8-0d45f75bed2c">

```
  Benchmark Results Summary    

duration phase no difference [-82ms to 137ms]
renderEnd phase estimated improvement -2ms [-4ms to 0ms] OR -3.32% [-5.93% to -0.73%]
render1000Items1End phase estimated improvement -49ms [-62ms to -43ms] OR -3.99% [-4.97% to -3.43%]
clearItems1End phase estimated improvement -3ms [-5ms to -1ms] OR -2.59% [-4.52% to -0.84%]
render1000Items2End phase estimated improvement -23ms [-30ms to -15ms] OR -2.34% [-3.06% to -1.56%]
clearItems2End phase estimated regression +2ms [1ms to 7ms] OR +4.53% [1.12% to 13.86%]
render5000Items1End phase estimated improvement -106ms [-137ms to -76ms] OR -2.32% [-3% to -1.68%]
clearManyItems1End phase estimated improvement -6ms [-9ms to -3ms] OR -2.59% [-3.73% to -1.24%]
render5000Items2End phase estimated regression +463ms [169ms to 506ms] OR +11.52% [4.22% to 12.6%]
clearManyItems2End phase estimated improvement -7ms [-10ms to -4ms] OR -3.18% [-4.26% to -1.71%]
render1000Items3End phase estimated improvement -34ms [-50ms to -19ms] OR -4.14% [-6.08% to -2.32%]
append1000Items1End phase estimated improvement -33ms [-39ms to -22ms] OR -3.57% [-4.21% to -2.35%]
append1000Items2End phase estimated improvement -33ms [-39ms to -23ms] OR -3.53% [-4.12% to -2.45%]
updateEvery10thItem1End phase estimated improvement -4ms [-8ms to -2ms] OR -2.45% [-4.67% to -1.07%]
updateEvery10thItem2End phase no difference [0ms to 1ms]
selectFirstRow1End phase estimated regression +16ms [6ms to 17ms] OR +11.72% [4.28% to 12.55%]
selectSecondRow1End phase estimated regression +5ms [1ms to 9ms] OR +3.94% [1.01% to 6.46%]
removeFirstRow1End phase estimated regression +16ms [14ms to 17ms] OR +7.58% [6.26% to 7.84%]
removeSecondRow1End phase estimated regression +11ms [3ms to 15ms] OR +5.01% [1.32% to 6.96%]
swapRows1End phase estimated improvement -1ms [-2ms to 0ms] OR -0.37% [-1.12% to -0.04%]
swapRows2End phase no difference [-3ms to 0ms]
clearItems4End phase estimated regression +10ms [6ms to 13ms] OR +7.08% [4.3% to 9.33%]
paint phase no difference [0ms to 0ms]
```

[artifact-1.pdf](https://github.com/glimmerjs/glimmer-vm/files/13762833/artifact-1.pdf)